/refactorFileLayout: Refactor Layout
Change filenames and move code around to make better sense from a higher-level. Add Amount type.
