+++ instruments.go
@@ -7,15 +7,17 @@ import (
 	"time"
 )
 
-type Currency uint64
-
-// NOTE:is this even needed?
-// func NewCurrency(in string) {
-// 	strconv.Itoa(in)
+// ...modelable? investmentable?
+// TODO: type marketable interface {
+// 	updateMetrics()
 // }
 
-func (c Currency) String() string {
-	str := strconv.Itoa(c)
+// Amount ... TODO
+type Amount uint64
+
+// Currency ... TODO
+func (c Amount) Currency() string {
+	str := strconv.Itoa(int(c))
 
 	b := bytes.NewBufferString(str)
 	numCommas := (b.Len() - 2) / 3
@@ -44,15 +46,15 @@ func (c Currency) String() string {
 // holds aggregate information regarding a financial asset, it is embeded into an Index or Benchmark.
 type Security struct {
 	Ticker                        string
-	NumTicks                      int
+	NumTicks                      uint
+	AvgVolume, AvgPrice           Amount
 	LastPrice, MaxPrice, MinPrice datedMetric
 	MaxVolume, MinVolume          datedMetric
 	BuyPrice, SellPrice           datedMetric
-	AvgPrice, AvgVolume           Currency, float64
 }
 
 // NewSecurity instantiates a new security from Tick data.
-func NewSecurity(tick Tick) (newSecurity *Security) {
+func NewSecurity(tick Tick) *Security {
 	firstPrice := datedMetric{Amount: tick.Price, Date: tick.Datetime}
 	firstVolume := datedMetric{Amount: tick.Volume, Date: tick.Datetime}
 	return &Security{
@@ -60,37 +62,36 @@ func NewSecurity(tick Tick) (newSecurity *Security) {
 		LastPrice: firstPrice, BuyPrice: firstPrice,
 		AvgPrice: tick.Price, AvgVolume: tick.Volume,
 		MaxPrice: firstPrice, MinPrice: firstPrice,
-		MaxVolume: firstVolume, MinVolume: firstVolume}
+		MaxVolume: firstVolume, MinVolume: firstVolume,
+	}
 }
 
+// TODO updateMetrics ...
 func (security *Security) updateMetrics(tick Tick) {
-	// TODO
+
 }
 
 type datedMetric struct {
-	Amount Currency
+	Amount Amount
 	Date   time.Time
 }
 
 // Position structs refer the holding of a financial asset.
 type Position struct {
 	Ticker                        string
-	Volume                        float64
-	NumTicks                      int
-	AvgPrice                      Currency
+	Volume                        Amount
+	NumTicks                      uint
+	AvgPrice                      Amount
 	LastPrice, MaxPrice, MinPrice datedMetric
 	BuyPrice, SellPrice           datedMetric
 }
 
-// TODO: favor updateMetrics instad
-func (pos *Position) update(tick Tick) {
-	pos.LastPrice = datedMetric{tick.Price, tick.Datetime}
-}
-
 func (pos *Position) updateMetrics(tick Tick) (ok bool) {
-	pos.AvgPrice = func() Currency {
-		numerator := (pos.AvgPrice * pos.NumTicks) + tick.Price
-		return numerator / (pos.NumTicks + 1)
+	pos.LastPrice = datedMetric{tick.Price, tick.Datetime}
+
+	pos.AvgPrice = func() Amount {
+		numerator := (pos.AvgPrice * Amount(pos.NumTicks)) + tick.Price
+		return numerator / (Amount(pos.NumTicks) + 1)
 	}()
 	pos.NumTicks++
 
@@ -110,66 +111,18 @@ func (pos *Position) updateMetrics(tick Tick) (ok bool) {
 	return true
 }
 
-// TODO: Move this
-func (pos *Position) sellShares(order *Order, amtToSell float64) *Position {
-
-	soldPos := func() *Position {
-		posSold := *pos
-		posSold.Volume = amtToSell
-		posSold.SellPrice = datedMetric{order.Price, order.Datetime}
-		return &posSold
-	}()
-	// Update active volume for pos
-	pos.Volume = pos.Volume - amtToSell
-
-	return soldPos
-}
-
 // Tick structs holds information about a financial asset at a specific point in time.
 type Tick struct {
 	Ticker   string
-	Price    float64
-	Volume   float64
-	BidSize  float64
-	AskSize  float64
+	Price    Amount
+	Volume   Amount
+	BidSize  Amount
+	AskSize  Amount
 	Datetime time.Time
 }
 
-// Order structs hold information referring to the details of the execution of a financial asset transaction.
-type Order struct {
-	// it's either buy or sell
-	Buy      bool
-	Status   OrderStatus
-	Logic    TradeLogic
-	Ticker   string
-	Price    float64
-	Volume   float64
-	Datetime time.Time
-}
-
-// OrderStatus variables refer to a status of an order's execution.
-type OrderStatus int
-
-const (
-	open OrderStatus = iota // 0
-	completed
-	canceled
-	expired // 3
-)
-
-// TradeLogic is used to identify when the order should be executed.
-type TradeLogic int
-
-const (
-	market TradeLogic = iota // 0
-	limit
-	stopLimit
-	stopLoss
-	day // 4
-)
-
-// NOTE: is this really needed?
 // Kwarg struct allows for add'l args/attrs to a class or func.
+// NOTE: is this really needed?
 type Kwarg struct {
 	name  string
 	value interface{}
